Moved cli.py from core to command_line.  Moved settings to top level folder. Moved parser functions from ashcalc into cli.  Moved isopach reading code to isopach_file.py.
